feat(submission): refactor submission handling by moving related functions to a new submission module; flatten evaluation parralelism
